adding authorization with access token

+++ src/api/notes/handler.js
@@ -16,8 +16,10 @@ class NotesHandler {
   async postNoteHandler(request, h) {
     this._validator.validateNotePayload(request.payload);
     const { title = 'untitled', body, tags } = request.payload;
-
-    const noteId = await this._service.addNote({ title, body, tags });
+    const { id: credentialId } = request.auth.credentials;
+    const noteId = await this._service.addNote({
+      title, body, tags, owner: credentialId,
+    });
 
     const response = h.response({
       status: 'success',
@@ -30,8 +32,9 @@ class NotesHandler {
     return response;
   }
 
-  async getNotesHandler() {
-    const notes = await this._service.getNotes();
+  async getNotesHandler(request) {
+    const { id: credentialId } = request.auth.credentials;
+    const notes = await this._service.getNotes(credentialId);
     return {
       status: 'success',
       data: {
@@ -42,6 +45,9 @@ class NotesHandler {
 
   async getNoteByIdHandler(request, h) {
     const { id } = request.params;
+    const { id: credentialId } = request.auth.credentials;
+    
+    await this._service.verifyNoteOwner(id, credentialId);
     const note = await this._service.getNoteById(id);
     return {
       status: 'success',
@@ -51,9 +57,11 @@ class NotesHandler {
     };
   }
 
-  putNoteByIdHandler(request, h) {
+  async putNoteByIdHandler(request, h) {
     this._validator.validateNotePayload(request.payload);
     const { id } = request.params;
+    const { id: credentialId} = request.auth.credentials;
+    await this._service.verifyNoteOwner(id, credentialId);
 
     this._service.editNoteById(id, request.payload);
 
@@ -63,8 +71,10 @@ class NotesHandler {
     };
   }
 
-  deleteNoteByIdHandler(request, h) {
+  async deleteNoteByIdHandler(request, h) {
     const { id } = request.params;
+    const { id: credentialId } = request.auth.credentials;
+    await this._service.verifyNoteOwner(id, credentialId);
     this._service.deleteNoteById(id);
 
     return {
